backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
aiosqlite==0.19.0
pydantic==2.5.0
websockets==12.0
aiohttp==3.9.1
python-dotenv==1.0.0
redis==5.0.1
Pillow==10.1.0
python-telegram-bot==20.7
near-api-py==0.25.0
web3==6.11.4
solana==0.30.0
pytezos==3.12.0